Fix build problem with gcc 4.6 on 64 bit
As reported in https://bugs.launchpad.net/ubuntu/+source/hamfax/+bug/770985 and
debugged by John Nogatch, gcc requires that the library has to be specified
after the module that references it.

Fix this by specifying the Qt libraries in LIBS, not in LDFLAGS and adjust the
test in bnc_try_4 as well. Also remove the empty argument to AC_TRY_LINK, that
seems to break newer autoconf versions.