Improve saving of images
When saving a file through File->Save and there is no valid image extension in
the file name, the file is not saved and there is no error message. This patch
uses PNG as default image format when there is no valid image extension and PNG
is support by the current Qt library. In case of any problem while saving a
file, a new warning message is displayed.