Implement FaxImage without Q3ScrollView
The Q3ScrollView provided a widget for the area inside the scroll bars, but the
Qt4 replacement class QScrollArea does not. Implement a new widget in
ImageWidget that draws itself with data from the QImage that is used for the
fax image data. With this approach, the changes to FaxImage are minimal: Use
ImageWidget as widget in the scroll area, update it through widget()->update()
and transform the mouse click coordinates from the scroll area widget to the
ImageWidget.