Avoid Qt3 support functions in various files
Using the Qt3 support library also enables additional functions in Qt4 classes.
As one step to move away from the Qt3 support library, convert most hamfax
files to avoid the Qt3 support functions. This is mostly trivial renaming. For
access to the config file, it is easier to implement functions with the Qt3
signatures, since they are accessed throughout the hamfax code.