Convert FaxWindow to avoid the Qt3 support functions
Without Qt3 support, the menus and tool bars have to be setup using
addAction().  The convenient approach in passing an integer from the menus to
the initTransmit() or initReception() slot has also disappeared, so split these
functions and connect the menu actions seperately.  It would also be a good
idea to split the endTransmit() and endReception() functions in a similar way,
but this can wait for later; add a TODO entry.

Menu entries are now referenced through QAction pointers, not through indexes.

And also apply the obvious renames and changed prototypes for the switch from
the Qt3 support functions.