Remove unused HF_QT_THREAD ifdefs
hamfax is a single threaded application. All I/O is started from the central
select()/poll() call in Qt and processed by the single thread. This makes it
much easier to debug and is fast enough. Remove the unused and untested ifdefs
that would start an additional thread for Alsa. This also makes the code more
readable.