Switch ALSA implementation from notifiers to poll model
Switch the ALSA implementation to use the poll()/select() loop in
QSocketNotifier. This allows the removal of the ALSA callback handlers and the
internal helper socket, and it follows the same idea as the other hamfax I/O
interfaces.

The only special thing is that ALSA may use more than one file descriptor and
the indicated I/O direction does not necessarily match the actual flow of sound
data. This is accomodated by creating one QSocketNotifier for each file
descriptor and each data flow direction that is requrested by ALSA, and using
snd_pcm_poll_descriptors_revents to determine the actual event.