Fix checkable option menu entries
After the port to Qt4, the checkable entries in the option menu could not be
changed. Fix this by using the checkable() signal. This already provides the
new state of the check, so the slot functions can be simplified and the
pointers to the QActions can be removed from the FaxWindow class.