Fix string termination for hamlib_parameters
The terminating 0 is missing for the hamlib_parameters. The string data is
copied to the vector, but the terminating 0 is missing (STL says that the
"last" iterator points _behind_ the last byte). Fix this by removing the
copy to the std::vector, the std::string can be accessed the same way, and it
has the correct string data.