Do not open hamlib rig for release only
The sound reception calls PTT::release in the final cleanup. This triggers
opening of the hamlib rig, only to release the PTT that has not been set
before.  Change this so that only if the rig is already open, the PTT is set,
and it makes sense to release it and close the hamlib rig.